requirements.txt
blinker==1.8.2
click==8.1.3
colorama==0.4.6
experta==1.9.4
filelock==3.14.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozendict==1.2
fsspec==2024.5.0
greenlet==3.0.3
gunicorn==20.1.0
importlib-metadata==4.11.3
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
MarkupSafe==2.1.1
mkl==2021.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
patsy==0.5.6
pgmpy==0.1.25
psycopg2==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
schema==0.6.7
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.30
statsmodels==0.14.1
sympy==1.12.1
tbb==2021.12.0
threadpoolctl==3.5.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.8.0